moved methods for media to view model
